+++ CHANGELOG.md
@@ -4,18 +4,28 @@
 
 ### Future
 #### NEW
-- ...
+- add inline asciimath support
+- rename to showdown-katex?
 
 #### FIX
-- become independent from jQuery
+- add inline latex examples
+- move `window.katex.config` to `window.katexLatex.config`?s
 
-### 0.3.0 2016-12-02
+### 0.3.0
 #### NEW
 - `katex.config` object used to configure katex displayMode, errorColor, and throwOnError
+- moved from gulp to npm scripts and rollup based build system
+- moved `dist/ASCIIMathTeXImg.js` to `src/asciimath-to-tex.js`, adopting it into the source of the project
+- treeshaking the extra functions in `asciimath-to-tex.js`
+- moved to es6 thanks to babel and rollup
+- added katex auto-render extension integration. it is not bundled along with katex-latex. so there is inline latex support now!
+- exports showdown extension function as default and defines `window.katexLatex` function
 
 #### FIXED
 - wrapping `<div>` was substituted with a `<span>` to allow for flexibility
 - demo now pretty prints normal code by using `bootmark`
+- no more jquery dependency!
+- better organized examples
 
 ### 0.2.0 2016-10-10
 #### NEW